feat: implement folder scanning and MHTML extraction
Add functionality to scan specified folders for MHTML files and 
extract relevant information such as URL and subject. Introduce 
priority folder handling to ensure important folders are scanned 
last. Implement a utility to generate unique Markdown file paths 
to avoid overwriting existing files.
