+++ folder_scanner.py
@@ -1,29 +1,7 @@
 import os
-import argparse
-from folder_scanner import scan_folders_and_create_md
-from markdown_utils import get_unique_md_path
+from file_operations import extract_info_from_mhtml
 
 PRIORITY_FOLDERS = ["@ Bads", "@ Other", "@ Weak", "@ Dead"]  # Folders to scan first
-IGNORE_FOLDERS = ["- Theory"]  # Folders to ignore
-
-def extract_info_from_mhtml(file_path):
-    """Extracts Content-Location (URL) and Subject from an MHTML file."""
-    try:
-        with open(file_path, 'r', encoding="utf-8", errors="ignore") as f:
-            content = f.read()
-
-        # Extract Content-Location (URL)
-        url_match = re.search(r'Content-Location:\s*(https?://[^\s]+)', content)
-        url = url_match.group(1) if url_match else "Unknown URL"
-
-        # Extract Subject
-        subject_match = re.search(r'Subject:\s*(.+)', content)
-        subject = subject_match.group(1).strip() if subject_match else "No Subject"
-
-        return url, subject
-    except Exception as e:
-        print(f"Error processing {file_path}: {e}")
-        return None, None
 
 def scan_folder(folder_path, md_file):
     """Scans a folder for MHTML files and writes to the markdown file."""
@@ -48,21 +26,6 @@ def scan_folder(folder_path, md_file):
                     md_file.write(f"[{url}] - {subject}\n")
                     print(f"Extracted: {url} - {subject}")
 
-def get_unique_md_path(output_md_path):
-    """Generates a unique markdown file path if the file already exists."""
-    if not os.path.exists(output_md_path):
-        return output_md_path
-    
-    base, ext = os.path.splitext(output_md_path)
-    counter = 1
-    new_output_md_path = f"{base}_{counter}{ext}"
-    
-    while os.path.exists(new_output_md_path):
-        counter += 1
-        new_output_md_path = f"{base}_{counter}{ext}"
-    
-    return new_output_md_path
-
 def scan_folders_and_create_md(root_folder, output_md_path):
     """Scans subfolders, prioritizes certain folders, and creates a Markdown file."""
     
@@ -81,17 +44,4 @@ def scan_folders_and_create_md(root_folder, output_md_path):
                 print(f"Scanning priority folder: {folder_path}")
                 scan_folder(folder_path, md_file)
 
-    print(f"Markdown file saved at: {output_md_path}")
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Scan MHTML files and extract information into a markdown file.")
-    parser.add_argument("root_folder", help="Path to the root folder to scan.")
-    parser.add_argument("output_md", nargs="?", help="Path to save the output markdown file.")
-    
-    args = parser.parse_args()
-    
-    # If output path is not provided, use the root folder with default filename
-    if args.output_md is None:
-        args.output_md = os.path.join(args.root_folder, "README.md")
-    
-    scan_folders_and_create_md(args.root_folder, args.output_md)
+    print(f"Markdown file saved at: {output_md_path}")
